requirements.txt
Flask>=2.1,<3.0
Flask-Cors>=3.0,<7.0
python-dotenv
pyodbc
google-generativeai